Move AggregateMessages into submodule lib
