+++ python/docs/graphframes.lib.rst
@@ -1,0 +1,10 @@
+graphframes.lib package
+=======================
+
+Collection of utilities usable with message aggregation
+-------------------------------------------------------
+
+.. automodule:: graphframes.lib
+    :members:
+    :undoc-members:
+    :inherited-members:
