+++ python/docs/graphframes.rst
@@ -9,6 +9,8 @@ Subpackages
 
     graphframes.examples
 
+    graphframes.lib
+
 Contents
 --------
