+++ python/graphframes/lib/__init__.py
@@ -1,0 +1,4 @@
+
+from .aggregate_messages import AggregateMessages
+
+__all__ = ['AggregateMessages']
